+++ simple-logger.js
@@ -1,0 +1,6 @@
+const simpleLogger = (req, res, next) => {
+    console.log(`Request ${req.method}:${req.url} ${res.statusCode}`)
+    next()
+}
+
+module.exports = simpleLogger