+++ VersionHistory.md
@@ -1,5 +1,9 @@
+0.4.4 [Oct 2020]
+- simple logger middleware
+- minor validation
+
 0.4.3 [Oct 2020]
-- attemp to write unit tests
+- attempt to write unit tests
 
 0.4.2 [Sep 2020]
 - fix an error when reading states delta?